extract globalSearchConfig variable so we can simplify config.js
